@@ -1,33 +1,28 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.2'
+gem 'rails', '~> 4.2.1'
 
-# Bundle edge Rails instead:
-# gem 'rails', :git => 'git://github.com/rails/rails.git'
+
+gem 'jquery-rails'
+
 
 gem 'sqlite3'
 gem 'libv8'
 gem 'therubyracer'
+gem 'protected_attributes'
 
 
 # Gems used only for assets and not required
 # in production environments by default.
 group :assets do
-  gem "less-rails" #Sprockets (what Rails 3.1 uses for its asset pipeline) supports LESS
-  gem 'less'
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
 
   # See https://github.com/sstephenson/execjs#readme for more supported runtimes
   # gem 'therubyracer'
-
+  gem 'sass-rails', '~> 5.0.3'
+  gem 'coffee-rails', '~> 4.1.0'
+  gem 'twitter-bootstrap-rails', '~> 3.2.0'
   gem 'uglifier', '>= 1.0.3'
-  gem 'twitter-bootstrap-rails', :git => 'http://github.com/seyhunak/twitter-bootstrap-rails.git'
 end
-
-gem 'jquery-rails'
-gem 'basecamp'
-gem 'xml-simple'
 
 # To use ActiveModel has_secure_password
 # gem 'bcrypt-ruby', '~> 3.0.0'
